update : - move the types to separate file - remove unused module - change endpoint uri on register
